Misc. unsigned warning fixes
Address a few more unsigned changes that don't fit in elsewhere.

+++ tools/tools-shared.c
@@ -113,7 +113,7 @@ int set_serial_attr(int fd, unsigned int baud)
 
 int write_to_tablet(int fd, const char *command)
 {
-	int len = 0;
+	unsigned long len = 0;
 
 	do {
 		int l;
